prevent neighbor list access out of bounds

+++ sph/include/sph/sph.cuh
@@ -11,29 +11,29 @@ namespace cuda
 {
 
 template<class Dataset>
-extern void computeDensity(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeDensity(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeIAD(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeIAD(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeMomentumEnergySTD(size_t, size_t, size_t, Dataset& d,
+extern void computeMomentumEnergySTD(size_t, size_t, int, Dataset& d,
                                      const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeXMass(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeXMass(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeVeDefGradh(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeVeDefGradh(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeIadDivvCurlv(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeIadDivvCurlv(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeAVswitches(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeAVswitches(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 template<class Dataset>
-extern void computeMomentumEnergy(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+extern void computeMomentumEnergy(size_t, size_t, int, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
 
 } // namespace cuda
 } // namespace sph
